backend/requirements.txt
alembic==1.16.3
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.39.3
botocore==1.39.3
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dlib @ file:///D:/face_attendance_app/backend/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.116.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python==4.12.0.88
passlib==1.7.4
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
